pyproject.toml
[project]
name = "productivity-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "channels>=4.3.2",
    "channels-redis>=4.3.0",
    "dj-rest-auth[with-social]>=7.0.1",
    "django>=5.2.8",
    "django-allauth>=65.13.1",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "faiss-cpu>=1.13.0",
    "hypercorn>=0.18.0",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.0",
    "langchain-openai>=1.1.6",
    "llama-cpp-python>=0.3.16",
    "psycopg2>=2.9.11",
    "redis>=7.1.0",
]
